+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [v1.18.0] - 2024-11-06
+
 ### Added
 
 * New tap action to pause/resume automatic photo cycling
@@ -15,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 * Skip the unsaved changes warning when there are no changes
+* Added device details to bug reports
 
 ## [v1.17.0] - 2024-10-27